chore: cargo fmt + clippy --all-targets -D warnings clean
Apply rustfmt across the workspace and fix or annotate every clippy lint
that surfaced under `--all-features --all-targets -- -D warnings`.
Most of the diff is mechanical formatter output; the substantive changes
are listed below.

Substantive fixes
-----------------
- spvirit-types: collapse nested `if let ... if val >=/<= ...` blocks in
  `update_alarm_from_value()` into `if let ... && ... ` chains.
- spvirit-codec/epics_decode.rs: drop a dead `let protocol = protocol;`
  rebinding.
- spvirit-codec/spvirit_encode.rs: `#[allow(clippy::too_many_arguments)]`
  on `encode_search_response`, `encode_client_connection_validation`,
  `encode_beacon` (signatures are PVA wire-format mandated; refactoring
  to a struct is a separate concern).
- spvirit-server/pva_server.rs: replace
  `let mut config = Default::default(); config.x = ...;` with a struct
  literal + `..Default::default()` spread.
- spvirit-server/types.rs: convert `RecordInstance::writable` from a
  match-with-arms-returning-bool into a `matches!(...)`.
- spvirit-server/simple_store.rs: convert `PvStore` impl's read-only
  methods (`has_pv`, `get_snapshot`, `get_descriptor`, `is_writable`,
  `list_pvs`) from `fn ... -> impl Future + Send` into the equivalent
  `async fn`.
- spvirit-tools/spvirit_list.rs: `for idx in 0..12 { guid[idx] = ... }`
  → `for (idx, slot) in guid.iter_mut().enumerate() { *slot = ... }`.
- spvirit-tools/spvirit_search.rs: `#[allow(clippy::type_complexity)]`
  on the CID cache HashMap field; `#[allow(clippy::enum_variant_names)]`
  on `SortMode { ByName, BySearchCount, ... }` (the prefix is the
  intentional public-facing naming).
- spvirit-tools/spvirit_dodeca.rs / spvirit_get_compare.rs: `#[allow(
  clippy::too_many_arguments)]` on a couple of pixel-drawing /
  protocol-reproduction helpers.
- spvirit-tools/tests/spvirit_nt_codec.rs: literal floats `3.1415` /
  `2.7182` → `std::f64::consts::PI` / `::E` (approx_constant).
- spvirit-tools/tests/interop/epics_base.rs:
  `reader.lines().flatten().collect()` →
  `reader.lines().map_while(Result::ok).collect()` to avoid the
  `lines_filter_map_ok` infinite-loop hazard.

Verification
------------
- `cargo fmt --check`: clean.
- `cargo clippy --workspace --all-features --all-targets -- -D warnings`:
  zero errors, zero warnings.
- `cargo test --workspace --no-fail-fast`: same set of pre-existing
  fixture failures as before (5 tests that spawn an external
  `spvirit_server` binary not present in the test environment); no new
  regressions.

+++ spvirit-server/examples/exotic_nt.rs
@@ -4,8 +4,8 @@ use spvirit_server::{
     DbCommonState, OutputMode, PvaServer, RecordData, RecordInstance, RecordType,
 };
 use spvirit_types::{
-    NdCodec, NdDimension, NtNdArray, NtPayload, NtScalar, NtTable, NtTableColumn,
-    ScalarArrayValue, ScalarValue,
+    NdCodec, NdDimension, NtNdArray, NtPayload, NtScalar, NtTable, NtTableColumn, ScalarArrayValue,
+    ScalarValue,
 };
 
 #[tokio::main]
@@ -35,7 +35,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 "Acquire".to_string(),
                 "Calibrate".to_string(),
             ];
-            mode_nt.display_description = format!("Mode={}", mode_nt.display_form_choices[mode_idx as usize]);
+            mode_nt.display_description =
+                format!("Mode={}", mode_nt.display_form_choices[mode_idx as usize]);
             store.put_nt("SIM:MODE", NtPayload::Scalar(mode_nt)).await;
 
             // NTTable with two columns that change over time.
@@ -64,7 +65,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
             // NTNDArray with a tiny 4x4 image.
             let pixels = (0..16)
-                .map(|i| (((i as i32 + tick as i32) % 16) * 16) as u8)
+                .map(|i| (((i + tick as i32) % 16) * 16) as u8)
                 .collect::<Vec<_>>();
             let ndarray_nt = NtNdArray {
                 value: ScalarArrayValue::U8(pixels),
@@ -98,7 +99,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 time_stamp: None,
                 display: None,
             };
-            store.put_nt("SIM:IMG", NtPayload::NdArray(ndarray_nt)).await;
+            store
+                .put_nt("SIM:IMG", NtPayload::NdArray(ndarray_nt))
+                .await;
 
             if let Some(snapshot) = store.get_nt("SIM:MODE").await {
                 println!("SIM:MODE => {snapshot:?}");
